% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solvents}
\alias{solvents}
\title{solvents}
\format{A data frame with variables:
cid PubChem ID (unique)
name Name
bp Boiling point in °C
solubility Water solubility in mg/l at 25°C
density Density in g/ml at 20°C}
\source{
\enumerate{
\item Kim S, Thiessen PA, Bolton EE, Chen J, Fu G, Gindulyte A, Han L, He J, He
S, Shoemaker BA, Wang J, Yu B, Zhang J, Bryant SH. PubChem Substance and
Compound databases. Nucleic Acids Res. 2016 Jan 4; 44(D1):D1202-13. Epub
2015 Sep 22 [PubMed PMID: 26400175] doi: 10.1093/nar/gkv951
}
}
\usage{
solvents
}
\description{
Chemical data for common solvents
}
\keyword{datasets}
